# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "CortaAi API",
    "description": "API completa do sistema CortaAi — agendamentos de barbearia.\n\n**Autenticação:** Firebase Bearer Token.\n\nVariáveis de ambiente:\n- `base_url`: `https://api.cortaai.shop`\n\n**Fluxo básico:**\n1. `POST /api/auth/email/login` → salvar `idToken`\n2. Usar `Authorization: Bearer {idToken}` nas rotas protegidas",
    "version": "1.0.0",
    "contact": {
      "name": "CortaAi",
      "email": "contato@cortaai.shop"
    }
  },
  "servers": [
    {
      "url": "https://api.cortaai.shop",
      "description": "Produção"
    },
    {
      "url": "http://localhost:8080",
      "description": "Local (Docker)"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Firebase ID Token"
      }
    },
    "schemas": {
...
      "CreatePaymentRequest": {
        "type": "object",
        "required": [
          "appointmentId"
        ],
        "properties": {
          "appointmentId": {
            "type": "string",
            "format": "uuid",
            "example": "{{appointment_id}}"
          },
          "paymentMethod": {
            "type": "string",
            "enum": ["PIX", "CREDIT_CARD", "LOCAL"],
            "example": "CREDIT_CARD",
            "description": "Método de pagamento escolhido pelo cliente"
          }
        }
      },
...
      "MpConnectResponse": {
        "type": "object",
        "properties": {
          "authorizationUrl": {
            "type": "string",
            "format": "uri",
            "example": "https://auth.mercadopago.com/authorization?client_id=..."
          }
        }
      },
...
    }
  },
...
  "paths": {
...
    "/api/payments/mp-connect": {
      "get": {
        "tags": [
          "06 - Pagamentos"
        ],
        "summary": "Iniciar OAuth Mercado Pago (barbeiro)",
        "description": "Retorna a URL de autorização do Mercado Pago. O barbeiro deve ser redirecionado para ela.",
        "operationId": "mpConnect",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "state",
            "in": "query",
            "required": true,
            "description": "UUID do barbeiro — será devolvido pelo MP no callback",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "example": "{{barber_id}}"
          }
        ],
        "responses": {
          "200": {
            "description": "URL de autorização gerada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MpConnectResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/payments/mp-callback": {
      "get": {
        "tags": [
          "06 - Pagamentos"
        ],
        "summary": "Callback OAuth Mercado Pago",
        "description": "Recebe o code do MP, troca por access_token, salva no user-service e redireciona o barbeiro para /barberHome?mpLinked=true.",
        "operationId": "mpCallback",
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Código de autorização devolvido pelo Mercado Pago"
          },
          {
            "name": "state",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "UUID do barbeiro passado no state do mp-connect"
          }
        ],
        "responses": {
          "302": {
            "description": "Redireciona para https://cortaai.shop/barberHome?mpLinked=true"
          },
          "400": {
            "description": "Código inválido ou estado ausente"
          }
        }
      }
    },
    "/api/payments/create": {
      "post": {
        "tags": [
          "06 - Pagamentos"
        ],
        "summary": "Criar pagamento",
        "operationId": "createPayment",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "X-User-Id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "example": "{{customer_id}}",
            "description": "UUID do usuário (não injetado pelo gateway, deve ser enviado pelo client)"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePaymentRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Pagamento criado"
          }
        }
      }
    },
    "/api/payments/{paymentId}": {
      "get": {
        "tags": [
          "06 - Pagamentos"
        ],
        "summary": "Buscar pagamento por ID",
        "operationId": "getPaymentById",
        "parameters": [
          {
            "name": "paymentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "example": "{{payment_id}}"
          }
        ],
        "responses": {
          "200": {
            "description": "Pagamento encontrado"
          }
        }
      }
    },
    "/api/payments/my-payments": {
      "get": {
        "tags": [
          "06 - Pagamentos"
        ],
        "summary": "Meus pagamentos",
        "operationId": "myPayments",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "X-User-Id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "example": "{{customer_id}}"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de pagamentos"
          }
        }
      }
    },
...
  }
}